+++ src/constants/app.ts
@@ -1,4 +1,4 @@
-export const APP_VERSION = '1.1.11-preview';
+export const APP_VERSION = '1.1.14';
 
 export const APP_IMAGE_HEADER_MAGIC = 0xe9;
 export const APP_DESCRIPTOR_MAGIC = 0xabcd5432;